requirements-test.txt
django>=1.5.1
coverage
coveralls
mock>=1.0.1
pytest>=2.5.2
pytest-django>=2.6
flake8>=2.1.0
pep8>=1.4.5
pylint>=0.27.0
pyflakes
mccabe
factory_boy>=2.3.1
tox>=1.7.0

# Additional test requirements go here
ipython

